% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eudysbiome.R
\docType{data}
\name{harmGenera}
\alias{harmGenera}
\title{Manually curated genera annotation table
A data frame containing 59 genera annotated as "harmful" and the harmful species included in these genera.}
\format{A data frame with 118 rows and 2 columns specifying for \code{Genus} and \code{Species}.}
\usage{
data(harmGenera)
}
\value{
A data frame of 59 harmful genera and species included.
}
\description{
Manually curated genera annotation table
A data frame containing 59 genera annotated as "harmful" and the harmful species included in these genera.
}
\keyword{datasets}

