% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eudysbiome.R
\docType{data}
\name{microCount}
\alias{microCount}
\title{Microbial count contingency table}
\format{A data frame with 2 rows and 2 variables}
\usage{
data(microCount)
}
\value{
A data frame of differential abundance counts in condition-impact couple.
}
\description{
A matrix containing the counts of differential microbe classified into each condition-eubiotic/dysbiotic impact couple. Rows represent the condition comparisons, columns represent the eubiotic and dysbiotic impacts:
}
\details{
\itemize{
  \item EI. eubiotic impact
  \item DI. dysbiotic impact
}

The table can be produced by \code{\link{microCount}} function.
}
\keyword{datasets}

