% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqDownload.R
\name{seqDownload}
\alias{seqDownload}
\title{Download SILVA SSU rRNA Sequences}
\usage{
seqDownload(file, version, destdir)
}
\arguments{
\item{file}{the name of the sequence file to be downloaded, default to \code{SILVA_123_SSURef_Nr99_tax_silva_trunc.fasta.gz}.}

\item{version}{a number or character to specify the version of the downloaded file, default to version \code{123}.}

\item{destdir}{a directory path where the downloaded file is saved, default to the current directory.}
}
\value{
The absolute path of the uncompressed downloaded file.
}
\description{
Download the SILVA Small subunit (16S/18S, SSU) ribosomal RNA (rRNA) reference (Ref) dataset for all three domains of life (Bacteria, Archaea and Eukarya) from \href{http://www.arb-silva.de/}{SILVA}, default to download a non-redundant (NR) SSU Ref dataset (v123) built by a dereplication of the full SSU Ref using a 99% identify criterion.
}
\details{
The Ref dataset is downloaded from SILVA Archive, please make sure the file you downloaded and the version you choose are listed in the archive.
}
\examples{
#download the non-redundant SILVA SSU Ref(v123) dataset to the current directory
library("eudysbiome")
seqDownload()
}
\seealso{
\code{\link{seqExtract}}
}

