% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqExtract.R
\name{seqExtract}
\alias{seqExtract}
\title{Extract Matched Sequences}
\usage{
seqExtract(seq.file, tax.file, as ="RNAStringSet", seqExtract.file)
}
\arguments{
\item{seq.file}{a sequence file where to extract the sequences.}

\item{tax.file}{a file with the headers of the extracted sequences. Note: the headers start behind '>' as shown in sequence file.}

\item{as}{a character vector indicating the type of object to return, including \code{"RNAStringSet"}, \code{"DNAStringSet"}, and \code{"AAStringSet"}, default to \code{"RNAStringSet"}, see \code{\link[Rsamtools]{scanFa}}.}

\item{seqExtract.file}{a sequence file where the extracted sequences are saved.}
}
\value{
the absolute path of \code{seqExtract.file}.
}
\description{
Extract the sequences matched by the given sequence headers.
}
\details{
The function was specially used to extract sequences from the SILVA SSU rRNA Ref sequence dataset downloaded by \code{\link{seqDownload}}, based on the taxonomy mapping file with the extracted taxonomic paths, constructed by the \code{\link{taxExtract}} based on the given microbes.
}
\examples{
#not excute, extract sequences matched with taxonomic paths in taxExract.tax file
#seqExtract(seq.file = "SILVA_123_SSURef_Nr99_tax_silva_trunc.fasta", tax.file = "taxExtract.tax", 
#as = "RNAStringSet", seqExtract.file = "seqExtract.fasta")
}
\seealso{
\code{\link{seqDownload}},\code{\link{taxExtract}}
}

