% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxDownload.R
\name{taxDownload}
\alias{taxDownload}
\title{Download SILVA Taxonomy Mapping File}
\usage{
taxDownload(file = "taxmap_slv_ssu_123.txt", version = 123, destdir = ".")
}
\arguments{
\item{file}{a taxonomy mapping file to be downloaded, default to \code{taxmap_slv_ssu_123.txt}.}

\item{version}{a number or character to specify the version of the downloaded file, default to version \code{123}.}

\item{destdir}{a directory path where the downloaded file is saved, default to the current directory.}
}
\value{
The absolute path of the downloaded file.
}
\description{
Download the SILVA taxonomy mapping file from \href{http://www.arb-silva.de/}{SILVA}, which maps each entry in the SILVA SSU Ref dataset to a taxonomic path containing the taxonomy rank designations from kingdom to species.
}
\details{
The tax mapping file should be consistent (same version) with the downloaded reference dataset by \code{\link{seqDownload}}. And also, you should make sure the file is listed in SILVA Archive.
}
\examples{
#download the taxonomy mapping file(v123) of SILVA SSU Ref dataset to the current directory
taxDownload()
}
\seealso{
\code{\link{taxExtract}}
}

