% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxExtract.R
\name{taxExtract}
\alias{taxExtract}
\title{Extract Matched Taxonomic Paths}
\usage{
taxExtract(tax.file, microbe, taxExtract.file)
}
\arguments{
\item{tax.file}{a taxonomy mapping file with taxonomic paths, downloaded by \code{\link{taxDownload}}.}

\item{microbe}{a character vector of selected microbes, classified at any taxonomic rank of kingdom, phylum, class, order, family, genus or species.}

\item{taxExtract.file}{a taxonomy file where the extracted taxonomic paths are saved.}
}
\value{
the absolute path of the \code{taxExtract.file}
}
\description{
Extract the taxonomic paths from the taxonomy mapping file by pattern matching with given microbes.
}
\examples{
#extract taxonomic paths matched to "Firmutes"
phyla = "Firmutes"
#not excute
#taxExtract(tax.file = "taxmap_slv_ssu_123.txt", microbe = phyla, taxExtract.file = "taxExtract.tax")

#extract taxonomic paths matched to "Lactobacillus" and "Bacteroides" genera
genera = c("Lactobacillus","Bacteroides")
#not excute
#taxExtract(tax.file = "taxmap_slv_ssu_123.txt", microbe = genera, taxExtract.file = "taxExtract.tax")
}
\seealso{
\code{\link{taxDownload}}, \code{\link{seqExtract}}
}

