% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eudysbiome.R
\docType{data}
\name{diffGenera}
\alias{diffGenera}
\title{Differential microbes in Genus-Species table
A data frame containing 10 differential genera and the species included, which was to be annotated as \code{"harmful"} or \code{"harmless"}.}
\format{A data frame with 26 rows and 2 columns specifying for \code{Genus} and \code{Species}.}
\usage{
data(diffGenera)
}
\value{
A data frame of 10 differential genera and species included.
}
\description{
Differential microbes in Genus-Species table
A data frame containing 10 differential genera and the species included, which was to be annotated as \code{"harmful"} or \code{"harmless"}.
}
\keyword{datasets}

