% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eudysbiome.R
\docType{data}
\name{harmGenera}
\alias{harmGenera}
\title{Manually curated genera annotation table
A data frame containing 235 genera annotated as "harmful" and the harmful species included in these genera.}
\format{A data frame with 851 rows and 3 columns specifying for \code{Genus} and \code{Species} and the references.}
\usage{
data(harmGenera)
}
\description{
Manually curated genera annotation table
A data frame containing 235 genera annotated as "harmful" and the harmful species included in these genera.
}
\keyword{datasets}

