% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eudysbiome.R
\docType{data}
\name{microDiff}
\alias{microDiff}
\title{Differential annotated genera with abundance variations among pair-wise condition comparisons}
\format{A list}
\usage{
data(microDiff)
}
\value{
A list of differential genera and annotations
}
\description{
A list containing: i) a data frame of 10 differentila genera with abundance differences among 3 condition comparisons, in which row represents the differential microbes and column represents the comparisons; ii) Genera annotations for the 10 differential genera; iii) pre-defined condition comparison names
}
\keyword{datasets}

